requirements.txt
fastapi
uvicorn
python-dotenv
reportlab
supabase
